+++ DiscordBot/commands/mud/stats.js
@@ -1,3 +1,4 @@
+const {DEBUG} = require('../../globals.js');
 const commando = require('discord.js-commando');
 const db = require('../../../dbhandler');
 
@@ -14,8 +15,9 @@ class StatsCommand extends commando.Command {
     async run(message, args) {
         db.getItem(message.member.id, 'players', (data) => this.getPlayer(data, message));
 
-        // delete stat command after saying it
-        message.delete();
+        // delete the user's command if not debugging
+        if (!DEBUG)
+            message.delete();
     }
 
     getPlayer(data, message) {
